separate preferred-color-group-list
