+++ catalystwan/endpoints/configuration/policy/list/preferred_color_group.py
@@ -5,17 +5,13 @@ from uuid import UUID
 
 from catalystwan.endpoints import APIEndpoints, delete, get, post, put
 from catalystwan.endpoints.configuration.policy.abstractions import PolicyListEndpoints
-from catalystwan.models.policy.lists import PreferredColorGroupList
-from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListInfo, PolicyListPreview
+from catalystwan.models.policy.list.preferred_color_group import (
+    PreferredColorGroupList,
+    PreferredColorGroupListEditPayload,
+    PreferredColorGroupListInfo,
+)
+from catalystwan.models.policy.policy_list import InfoTag, PolicyListId, PolicyListPreview
 from catalystwan.typed_list import DataSequence
-
-
-class PreferredColorGroupListEditPayload(PreferredColorGroupList, PolicyListId):
-    pass
-
-
-class PreferredColorGroupListInfo(PreferredColorGroupList, PolicyListInfo):
-    pass
 
 
 class ConfigurationPreferredColorGroupList(APIEndpoints, PolicyListEndpoints):
