+++ catalystwan/models/policy/__init__.py
@@ -23,6 +23,7 @@ from catalystwan.models.policy.list.local_domain import LocalDomainList
 from catalystwan.models.policy.list.mirror import MirrorList
 from catalystwan.models.policy.list.policer import PolicerList
 from catalystwan.models.policy.list.port import PortList
+from catalystwan.models.policy.list.preferred_color_group import PreferredColorGroupList
 
 from .centralized import CentralizedPolicy, TrafficDataDirection
 from .definition.access_control_list import AclPolicy
@@ -40,7 +41,6 @@ from .definition.traffic_data import TrafficDataPolicy
 from .definition.vpn_membership import VPNMembershipPolicy
 from .definition.zone_based_firewall import ZoneBasedFWPolicy
 from .lists import (
-    PreferredColorGroupList,
     PrefixList,
     ProtocolNameList,
     RegionList,
