+++ catalystwan/models/policy/lists.py
@@ -8,10 +8,7 @@ from pydantic import BaseModel, Field
 
 from catalystwan.models.common import InterfaceType, TLOCColor
 from catalystwan.models.policy.lists_entries import (
-    ColorGroupPreference,
     EncapType,
-    PathPreference,
-    PreferredColorGroupListEntry,
     PrefixListEntry,
     ProtocolNameListEntry,
     RegionListEntry,
@@ -130,33 +127,6 @@ class TLOCList(PolicyListBase):
     def add_tloc(self, tloc: IPv4Address, color: TLOCColor, encap: EncapType, preference: Optional[int] = None) -> None:
         _preference = str(preference) if preference is not None else None
         self.entries.append(TLOCListEntry(tloc=tloc, color=color, encap=encap, preference=_preference))
-
-
-class PreferredColorGroupList(PolicyListBase):
-    type: Literal["preferredColorGroup"] = "preferredColorGroup"
-    entries: List[PreferredColorGroupListEntry] = []
-
-    def assign_color_groups(
-        self,
-        primary: Tuple[Set[TLOCColor], PathPreference],
-        secondary: Optional[Tuple[Set[TLOCColor], PathPreference]] = None,
-        tertiary: Optional[Tuple[Set[TLOCColor], PathPreference]] = None,
-    ) -> PreferredColorGroupListEntry:
-        primary_preference = ColorGroupPreference.from_color_set_and_path(*primary)
-        secondary_preference = (
-            ColorGroupPreference.from_color_set_and_path(*secondary) if secondary is not None else None
-        )
-        tertiary_preference = ColorGroupPreference.from_color_set_and_path(*tertiary) if tertiary is not None else None
-        entry = PreferredColorGroupListEntry(
-            primary_preference=primary_preference,
-            secondary_preference=secondary_preference,
-            tertiary_preference=tertiary_preference,
-        )
-        if self.entries:
-            self.entries[0] = entry
-        else:
-            self.entries.append(entry)
-        return entry
 
 
 class PrefixList(PolicyListBase):
